AWS Lambda Rest: split the amazon-lambda-rest integration module
- into amazon-lambda-rest-resteasy-reactive,
amazon-lambda-rest-reactive-routes, amazon-lambda-rest-funqy and
amazon-lambda-rest-servlet
- to avoid errors/conflicts among HTTP frameworks
- there was a regression in the tests caused by
e2ae8180db5f4d80a27e7bef471998765ba476bd during upgrade to Vert.x 4.5.1
