+++ integration-tests/amazon-lambda-rest-funqy/src/main/resources/application.properties
