+++ integration-tests/amazon-lambda-rest-funqy/src/test/java/io/quarkus/it/amazon/lambda/rest/funqy/AmazonLambdaV1SimpleIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.it.amazon.lambda.rest.funqy;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class AmazonLambdaV1SimpleIT extends AmazonLambdaV1SimpleTestCase {
+}
