+++ integration-tests/amazon-lambda-rest-reactive-routes/src/test/java/io/quarkus/it/amazon/lambda/rest/reactive/routes/AmazonLambdaV1SimpleIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.it.amazon.lambda.rest.reactive.routes;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class AmazonLambdaV1SimpleIT extends AmazonLambdaV1SimpleTestCase {
+}
