+++ integration-tests/amazon-lambda-rest-resteasy-reactive/src/test/java/io/quarkus/it/amazon/lambda/rest/resteasy/reactive/AmazonLambdaV1SimpleIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.it.amazon.lambda.rest.resteasy.reactive;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class AmazonLambdaV1SimpleIT extends AmazonLambdaV1SimpleTestCase {
+}
