+++ integration-tests/amazon-lambda-rest-servlet/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+quarkus.lambda-http.enable-security=true
+quarkus.swagger-ui.always-include=true
