--- integration-tests/amazon-lambda-rest/src/main/java/io/quarkus/it/amazon/lambda/v1/FunqyGreeting.java
@@ -1,17 +1,0 @@
-package io.quarkus.it.amazon.lambda.v1;
-
-import io.quarkus.funqy.Funq;
-import io.smallrye.mutiny.Uni;
-
-public class FunqyGreeting {
-
-    @Funq
-    public String funqy(String name) {
-        return "Make it funqy " + name;
-    }
-
-    @Funq
-    public Uni<String> funqyAsync(String name) {
-        return Uni.createFrom().item(() -> "Make it funqy " + name);
-    }
-}
