--- integration-tests/amazon-lambda-rest/src/main/java/io/quarkus/it/amazon/lambda/v1/GreetingServlet.java
@@ -1,27 +1,0 @@
-package io.quarkus.it.amazon.lambda.v1;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@WebServlet(name = "ServletGreeting", urlPatterns = "/servlet/hello")
-public class GreetingServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setStatus(200);
-        resp.addHeader("Content-Type", "text/plain");
-        resp.getWriter().write("hello");
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String name = req.getReader().readLine();
-        resp.setStatus(200);
-        resp.addHeader("Content-Type", "text/plain");
-        resp.getWriter().write("hello " + name);
-    }
-}
