--- integration-tests/amazon-lambda-rest/src/main/java/io/quarkus/it/amazon/lambda/v1/SecurityServlet.java
@@ -1,19 +1,0 @@
-package io.quarkus.it.amazon.lambda.v1;
-
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@WebServlet(name = "ServletSecurity", urlPatterns = "/servlet/security")
-public class SecurityServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setStatus(200);
-        resp.addHeader("Content-Type", "text/plain");
-        resp.getWriter().write(req.getUserPrincipal().getName());
-    }
-}
