--- integration-tests/amazon-lambda-rest/src/test/java/io/quarkus/it/amazon/lambda/AmazonLambdaV1SimpleIT.java
@@ -1,7 +1,0 @@
-package io.quarkus.it.amazon.lambda;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class AmazonLambdaV1SimpleIT extends AmazonLambdaV1SimpleTestCase {
-}
